Use nancy directly as a Go tool to avoid Docker dependency (#492)
Initially the Docker approach was attractive as it kept a nice separation of the nancy environment used to do Go code vulnerability scans. The build uses lots of Go tools directly already so it seems simpler to do the same with nancy too. This should make it easier to use a local exclude file checked into the repository if necessary, make it easier to keep the nancy version up-to-date, and remove the need for Docker outside of the scenario tests.

Signed-off-by: Mark S. Lewis <mark_lewis@uk.ibm.com>