Use only OSV-Scanner to scan Go code (#785)
OSV-Scanner uses Govulncheck internally so there is little benefit to
running both. Nancy does not perform call analysis when scanning so is
prone to false positives compared to OSV-Scanner (and Govulncheck).

This change removes scheduled scans of Go code using Govulncheck and
Nancy. Only OSV-Scanner is used for Go code. Scheduled scanning of the
latest code is added, in addition to scanning of the latest release.

Signed-off-by: Mark S. Lewis <Mark.S.Lewis@outlook.com>